Remove LedIndicator class
Remove LedIndicator class because of the uC switching (Firebettle to Wemos D1 mini)
Both class file are deleted and all the implementation of the LedIndicator class in the other classes are deleted too.
Note that the call of the killLoop() method has been replaced by ESP.restart() which forces the uC to reboot.